// src/types/themeTypes.ts
export interface themeObject {
  black: string;
  white: string;
  grey: string;
  lightGrey: string;
  secondaryBlack: string;
  lightOrange: string;
  orange: string;
}

export interface themeType {
  theme: {
    black: string;
    white: string;
    grey: string;
    lightGrey: string;
    secondaryBlack: string;
    lightOrange: string;
    orange: string;
  };
}